requirements.txt
olefile==0.44
Pillow==4.2.1
pycryptodome==3.4.6
